Ref move onSubmit function into ContactForm
